+++ src/main/java/edu/eci/cvds/services/BibliotecaServices.java
@@ -12,7 +12,7 @@ public interface BibliotecaServices {
         
         public Recurso consultarRecurso(int id) throws BibliotecaException;
         
-        public void updateRecurso(Long id, String nuevoEstado) throws BibliotecaException;
+        public void updateRecurso(int id, String nuevoEstado) throws BibliotecaException;
